pyproject.toml
[project]
name = "transformer-family"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "ipykernel>=7.2.0",
    "numpy>=2.4.2",
    "tiktoken>=0.12.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
